simplerllm_img_desktop_app/backend/requirements.txt
flask>=3.0
flask-cors>=4.0
python-dotenv>=1.0
simplerllm
Pillow>=10.0.0
